Named_PipeCreateScript
This PS script will create a named pipe for inter-process communication within the context of PowerShell. It allows one PowerShell script to write messages to the pipe and another to read from it. However, it doesn't interact directly with Windows services or applications outside of PowerShell. It operates purely within the PowerShell environment.